// paintPal-ui/src/app/services/models/page-response-painting-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PaintingResponse } from '../models/painting-response';
export interface PageResponsePaintingResponse {
  content?: Array<PaintingResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}