// paintPal-ui/src/app/services/models/painting-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface PaintingResponse {
  archived?: boolean;
  artist?: string;
  doi?: string;
  id?: number;
  image?: string;
  info?: string;
  owner?: string;
  rating?: number;
  shareable?: boolean;
  title?: string;
}